moving classes to separate files to improve future cleanup
