+++ src/WebJobs.Script.Grpc/Channel/IInvocationMessageDispatcherFactory.cs
@@ -1,0 +1,9 @@
+﻿// Copyright (c) .NET Foundation. All rights reserved.
+// Licensed under the MIT License. See License.txt in the project root for license information.
+
+namespace Microsoft.Azure.WebJobs.Script.Grpc;
+
+internal interface IInvocationMessageDispatcherFactory
+{
+    IInvocationMessageDispatcher Create(string invocationId);
+}
